refactor: Do the same thing with `ReplayGainParser`
- In the future, we may want to change up what we expose in `ReplayGainParser`, for example: `trackGain` (prev `gain`), `trackPeak`, `albumGain`, and `albumPeak`.
