+++ android/src/main/java/com/cyanchill/missingcore/metadataretriever/modules/ReplayGainParser.kt
@@ -1,4 +1,4 @@
-package com.cyanchill.missingcore.metadataretriever.utils
+package com.cyanchill.missingcore.metadataretriever.modules
 
 import androidx.annotation.OptIn
 import androidx.media3.common.Metadata
@@ -8,11 +8,21 @@ import androidx.media3.extractor.metadata.vorbis.VorbisComment
 import androidx.media3.extractor.mp3.Mp3InfoReplayGain
 
 @OptIn(UnstableApi::class)
-class ReplayGainParser(entry: Metadata.Entry) {
+class ReplayGainParser(metadataList: List<Metadata>) {
   var gain: Float? = null
 
   init {
-    // Extract track gain based on the class.
+    for (metadata in metadataList) {
+      val numEntries = metadata.length()
+      for (i in 0 until numEntries) {
+        parseMetadataEntry(metadata[i])
+        if (gain != null) break
+      }
+      if (gain != null) break
+    }
+  }
+
+  private fun parseMetadataEntry(entry: Metadata.Entry) {
     gain = when (entry) {
       is TextInformationFrame -> handleTextInformationFrame(entry)
       is Mp3InfoReplayGain -> handleMp3InfoReplayGain(entry)
@@ -21,6 +31,7 @@ class ReplayGainParser(entry: Metadata.Entry) {
     }
   }
 
+  //#region [ReplayGain Containers]
   private fun handleTextInformationFrame(frame: TextInformationFrame): Float? =
     when (frame.description?.uppercase()) {
       "REPLAYGAIN_TRACK_GAIN" -> frame.values.firstOrNull()?.parseReplayGainAdjustment()
@@ -41,8 +52,11 @@ class ReplayGainParser(entry: Metadata.Entry) {
       "R128_TRACK_GAIN" -> comment.value.parseReplayGainAdjustment()?.div(256f)
       else -> null
     }
+  //#endregion
 
+  //#region [Internal Overloads]
   /** Some replay gain tags include "dB" in the string. */
   private fun String.parseReplayGainAdjustment() =
     replace(Regex("[^\\d.-]"), "").toFloatOrNull()
+  //#endregion
 }
